+++ app/src/main/java/et/tv7/taevastv7/activities/MainActivity.java
@@ -35,6 +35,7 @@ import et.tv7.taevastv7.fragments.ProgramInfoFragment;
 import et.tv7.taevastv7.fragments.SearchFragment;
 import et.tv7.taevastv7.fragments.SearchResultFragment;
 import et.tv7.taevastv7.fragments.SeriesFragment;
+import et.tv7.taevastv7.fragments.SeriesInfoFragment;
 import et.tv7.taevastv7.fragments.TvMainFragment;
 import et.tv7.taevastv7.fragments.TvPlayerFragment;
 import et.tv7.taevastv7.helpers.GuideItem;
@@ -164,6 +165,10 @@ public class MainActivity extends AppCompatActivity implements ArchiveDataLoaded
                     // Program info fragment visible
                     return ((ProgramInfoFragment) fragment).onKeyDown(keyCode, events);
                 }
+                else if (fragment instanceof SeriesInfoFragment) {
+                    // Series info fragment visible
+                    return ((SeriesInfoFragment) fragment).onKeyDown(keyCode, events);
+                }
                 else if (fragment instanceof CategoriesFragment) {
                     // Categories fragment visible
                     return ((CategoriesFragment) fragment).onKeyDown(keyCode, events);
@@ -258,6 +263,8 @@ public class MainActivity extends AppCompatActivity implements ArchiveDataLoaded
                     else {
                         this.addGuideData(guideData);
 
+                        archiveViewModel.initializeSeriesData(guideViewModel.getGuide());
+
                         this.prepareUi();
                         Utils.toPage(TV_MAIN_FRAGMENT, this, false, false, null);
                     }
@@ -315,25 +322,25 @@ public class MainActivity extends AppCompatActivity implements ArchiveDataLoaded
             JSONObject obj = guideData.getJSONObject(i);
             if (obj != null) {
                 GuideItem g = new GuideItem(
-                    Utils.getJsonStringValue(obj, TIME),
-                    Utils.getJsonStringValue(obj, END_TIME),
-                    Utils.getJsonStringValue(obj, IMAGE_PATH),
-                    Utils.getJsonStringValue(obj, CAPTION),
-                    Utils.getJsonStringValue(obj, START_END_TIME),
-                    Utils.getJsonStringValue(obj, START_DATE),
-                    Utils.getJsonStringValue(obj, END_DATE),
-                    Utils.getJsonStringValue(obj, FORMATTED_START_TIME),
-                    Utils.getJsonStringValue(obj, FORMATTED_END_TIME),
-                    Utils.getJsonStringValue(obj, BROADCAST_DATE),
-                    Utils.getJsonStringValue(obj, BROADCAST_DATE_TIME),
-                    Utils.getJsonStringValue(obj, DURATION),
-                    Utils.getJsonStringValue(obj, SERIES),
-                    Utils.getJsonStringValue(obj, NAME),
-                    Utils.getJsonIntValue(obj, SID),
-                    Utils.getJsonIntValue(obj, EPISODE_NUMBER),
-                    Utils.getJsonIntValue(obj, IS_VISIBLE_ON_VOD),
-                    Utils.getJsonStringValue(obj, SERIES_AND_NAME),
-                    Utils.isStartDateToday(Utils.getJsonStringValue(obj, TIME)));
+                        Utils.getJsonStringValue(obj, TIME),
+                        Utils.getJsonStringValue(obj, END_TIME),
+                        Utils.getJsonStringValue(obj, IMAGE_PATH),
+                        Utils.getJsonStringValue(obj, CAPTION),
+                        Utils.getJsonStringValue(obj, START_END_TIME),
+                        Utils.getJsonStringValue(obj, START_DATE),
+                        Utils.getJsonStringValue(obj, END_DATE),
+                        Utils.getJsonStringValue(obj, FORMATTED_START_TIME),
+                        Utils.getJsonStringValue(obj, FORMATTED_END_TIME),
+                        Utils.getJsonStringValue(obj, BROADCAST_DATE),
+                        Utils.getJsonStringValue(obj, BROADCAST_DATE_TIME),
+                        Utils.getJsonStringValue(obj, DURATION),
+                        Utils.getJsonStringValue(obj, SERIES),
+                        Utils.getJsonStringValue(obj, NAME),
+                        Utils.getJsonIntValue(obj, SID),
+                        Utils.getJsonIntValue(obj, EPISODE_NUMBER),
+                        Utils.getJsonIntValue(obj, IS_VISIBLE_ON_VOD),
+                        Utils.getJsonStringValue(obj, SERIES_AND_NAME),
+                        Utils.isStartDateToday(Utils.getJsonStringValue(obj, TIME)));
 
                 guideViewModel.addItemToGuide(g);
             }